requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

# LLM inference
transformers==4.41.1
peft==0.10.0
bitsandbytes==0.43.1
accelerate==0.27.2
datasets==2.18.0

# Stacks & Clarity
# Handled by Node.js
